pyproject.toml
[tool.poetry]
name = "ot-harjoitustyo"
version = "0.1.0"
description = ""
authors = ["Janne <48507101+janneko01@users.noreply.github.com>"]
readme = "README.md"
packages = [{include = "ot_harjoitustyo"}]

[tool.poetry.dependencies]
python = "^3.8"
fpdf2 = "^2.6.0"
invoke = "^1.7.3"
python-dotenv = "^0.21.0"


[tool.poetry.group.dev.dependencies]
pytest = "^7.2.0"
pylint = "^2.15.6"
autopep8 = "^2.0.0"
coverage = "^7.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
